langgraph-convAgent/pyproject.toml
[project]
name = "lg-convagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain[google-genai]>=1.2.3",
    "langgraph>=1.0.5",
    "langgraph-cli[inmem]>=0.4.11",
    "python-dotenv>=1.2.1",
]
